moving all known keys to type
