+++ lib/types/index.ts
@@ -3,3 +3,8 @@ export * from './ITokenStore';
 export * from './ILogger';
 export * from './errors';
 export * from './common';
+export * from './DriveTrainType';
+
+// Re-export telematic keys
+export * from './TelematicKeys';
+export type { TelematicKeyDescriptor, EssentialKey } from './TelematicKeys';
